refactor folder strcture for design model
