--- software artifacts/domain model/ExcelSyncingService_SSD.puml
@@ -1,36 +1,0 @@
-@startuml
-
-使用者 -> 前端: 上傳Excel空白表單
-使用者 -> 前端: 設定資料綁定
-前端 -> 前端: 同步iSLM表格
-前端 -> 後端: 儲存Excel空白表單
-前端 -> 後端: JSON綁定物件
-
-後端 --> 前端: 返回Excel空白表單
-後端 --> 前端: 返回JSON綁定物件
-
-
-loop 使用iSLM
-  alt iSLM表格編輯
-    使用者 -> 前端: 編輯iSLM表格
-    前端 -> 前端: 資料驗證
-    前端 -> 後端: 欄位資料
-  else Excel表單編輯
-    loop
-      使用者 -> 前端: 編輯表單
-      後端 --> 前端: 返回JSON綁定物件
-      前端 -> 前端: JSON驗證及更新已綁定值
-      前端 -> 前端: 同步iSLM表格
-      前端 -> 後端: iSLM欄位值更新
-      前端 -> 後端: 儲存Excel表
-    end
-  else 更新JSON綁定物件
-    loop 
-      使用者 -> 前端: 編輯綁定
-      前端 -> 前端: 資料驗證
-      前端 -> 後端: JSON綁定物件
-    end
-  end
-end
-
-@enduml
