+++ src/app/service/EjectorService/EjectorValueBinding.service.ts
@@ -1,4 +1,4 @@
-import { ValueBindingService } from "./valueBinding.service";
+import { ValueBindingService } from "../ValueBinding.service";
 import { Injectable } from "@angular/core";
 
 @Injectable({providedIn: 'root'})
