+++ src/app/service/EjectorService/Properties/ejector.ts
