+++ src/app/service/PairFormService/PairFormBinding.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { ValueBindingService } from "./valueBinding.service";
+import { ValueBindingService } from "../ValueBinding.service";
 
 @Injectable({providedIn: 'root'})
 export class PairFormBindingService extends ValueBindingService {
