+++ src/app/service/ValueSync.service.ts
