+++ src/app/service/Window.service.ts
